Move CommentPanel to wagtail.admin.panels.comment_panel
